initial reorg to enable langs other than php
Summary: lex-pass 1.0 was very language agnostic and worked on token stream
instead of ast.  lex-pass 2.0 moved to ast which is of course a better
structure for manipulation, but it was very tied to php.  this reorg aims to
eventually make src/Lang/Js etc possible.

we might add some wrapper imports to cut down on imports Transfs need to do..

Reviewed By: srush

Test Plan: compiles

Revert Plan: ok