lex-pass rename-func/func-kill-arg transformers
Summary: lex-pass is best for crazily heady transformations, but
these down-to-earth changes are fine too.  and these ones are good
and reusable examples for the public repo.

Reviewed By: schrockn, eletuchy

Test Plan: ran func-kill-arg on an fb task for schrockn with n=2 on
func in hundreds of places with 1-3 args.  tested rename-func on
that func as well.

Revert Plan: ok