Fixes #1; colon notation for "if" control structures.
Added some things to help with testing.
I've used GenericPretty to display AST tree structures in debugging.
Unfortunately it requires "deriving Generic" and GHC.Generics.
We're already "deriving (Data, Typeable)" which seems to be similar
but not as GHC-only so probably better?  Have to look into.
Also GenericPretty makes things too wide.  I played around with an
alternative src/Test/PrettyPrint.hs but gave up on it for now..

Also tested by running dump-as-id on wordpress, which now works until
there's a "foreach .. :".  So I guess that's next.