invariants for everyone: NonEmpty lists in the AST
some parts of the AST have a list structure that cannot be empty.  this
is now an invariant in the type system.  there's a NonEmpty library on
hackage, but it's small and doesn't currently seem into my grand vision
of qualified importing to possibly have a lot of Data.List functionality.
or maybe that's stupid.  we'll see what happens.

also some cleanup here and trying to minimize the "Common" modules.

also pulled out WS into a module (long overdue).